requirements.txt
streamlit==1.32.0
flask==3.0.2
flask-sqlalchemy==3.1.1
pillow==10.2.0
python-dateutil==2.8.2 